chore: Add AddressEntity and update AdopterEntity to include address relationship
This commit adds a new entity called AddressEntity, which represents the address of an adopter. It includes columns for city and uf (state). The AdopterEntity is also updated to include a one-to-one relationship with the AddressEntity, allowing an adopter to have an associated address. The relationship is defined using the @OneToOne decorator and the @JoinColumn decorator specifies the field used to join the tables.

Refactor the Adopter type to use the AddressEntity instead of a string for the address field.